refactor(Search): wrap 8 SourceIndexer conformances + IndexerRegistry under Search; rename file
Continues #378 (SourceIndexer foundation). The 8 file-scope indexer
conformances + IndexerRegistry now live under Search. Every type in
Search/SourceIndexer.swift is now wrapped, so the file renames to
Search+SourceIndexer.swift per the <Namespace>+<Type>.swift rule.

## Renames

| Before | After |
|---|---|
| AppleDocsIndexer        | Search.AppleDocsIndexer |
| HIGIndexer              | Search.HIGIndexer |
| SwiftEvolutionIndexer   | Search.SwiftEvolutionIndexer |
| AppleArchiveIndexer     | Search.AppleArchiveIndexer |
| SwiftBookIndexer        | Search.SwiftBookIndexer |
| SwiftOrgIndexer         | Search.SwiftOrgIndexer |
| PackagesIndexer         | Search.PackagesIndexer |
| IndexerRegistry         | Search.IndexerRegistry |

## File rename

| Before | After |
|---|---|
| Search/SourceIndexer.swift | Search/Search+SourceIndexer.swift |
                              (dominant: SourceIndexer protocol;
                              carries 8 conformances, IndexerRegistry,
                              SourceItem, ExtractedContent companions)

## Verification

- xcrun swift build clean.
- xcrun swift test: 1300/1300 passing.
