+++ Packages/Sources/Shared/Constants/Shared.Constants.swift
@@ -1,5 +1,4 @@
 import Foundation
-import SharedConstants
 
 // MARK: - Cupertino Constants
